% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello()
}
\value{
A \code{character(1)} with the word "hello".
}
\description{
This function is just an example function to pass R CMD check.
}
\examples{
## Say hello ^_^
hello()
}
