% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIGDWorkflowDemoBioC2024-pkg.R
\docType{package}
\name{scIGDWorkflowDemoBioC2024-pkg}
\alias{scIGDWorkflowDemoBioC2024}
\alias{scIGDWorkflowDemoBioC2024-package}
\alias{scIGDWorkflowDemoBioC2024-pkg}
\title{scIGDWorkflowDemoBioC2024}
\description{
This package provides the material for the scIGD workflow demo at BioC2024. The workflow comprises two software (scIGD - Snakemake; SingleCellAlleleExperiment - R/Bioconductor) and one data (scaeData - R/ExperimentHub) packages. I showcase all three tools, emphasizing the utilization of SingleCellAlleleExperiment and its functionalities on one of the example datasets available in scaeData, for exploratory and downstream analysis across the data layers offered by the data structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ahmadalajami.github.io/scIGDWorkflowDemoBioC2024/}
  \item \url{ghcr.io/ahmadalajami/scigdworkflowdemobioc2024}
  \item Report bugs at \url{https://github.com/ahmadalajami/scIGDWorkflowDemoBioC2024/issues/new/choose}
}

}
\author{
\strong{Maintainer}: Ahmad Al Ajami \email{alajami@med.uni-frankfurt.de} (\href{https://orcid.org/0009-0006-5615-7447}{ORCID})

}
\keyword{internal}
